requirements.txt
torch
torchvision
streamlit
numpy
opencv-python
matplotlib
Pillow